[Enhancement] Query queue supports compute node (#17733)

+++ be/src/agent/agent_server.cpp
@@ -69,7 +69,7 @@ const uint32_t REPORT_RESOURCE_USAGE_WORKER_COUNT = 1;
 
 class AgentServer::Impl {
 public:
-    explicit Impl(ExecEnv* exec_env) : _exec_env(exec_env) {}
+    explicit Impl(ExecEnv* exec_env, bool is_compute_node) : _exec_env(exec_env), _is_compute_node(is_compute_node) {}
 
     ~Impl();
 
@@ -119,20 +119,24 @@ private:
     std::unique_ptr<ReportOlapTableTaskWorkerPool> _report_tablet_workers;
     std::unique_ptr<ReportWorkgroupTaskWorkerPool> _report_workgroup_workers;
     std::unique_ptr<ReportResourceUsageTaskWorkerPool> _report_resource_usage_workers;
+
+    // Compute node only need _report_resource_usage_workers.
+    const bool _is_compute_node;
 };
 
 void AgentServer::Impl::init_or_die() {
-    for (auto& path : _exec_env->store_paths()) {
-        try {
-            std::string dpp_download_path_str = path.path + DPP_PREFIX;
-            std::filesystem::path dpp_download_path(dpp_download_path_str);
-            if (std::filesystem::exists(dpp_download_path)) {
-                std::filesystem::remove_all(dpp_download_path);
+    if (!_is_compute_node) {
+        for (auto& path : _exec_env->store_paths()) {
+            try {
+                std::string dpp_download_path_str = path.path + DPP_PREFIX;
+                std::filesystem::path dpp_download_path(dpp_download_path_str);
+                if (std::filesystem::exists(dpp_download_path)) {
+                    std::filesystem::remove_all(dpp_download_path);
+                }
+            } catch (...) {
+                LOG(WARNING) << "std exception when remove dpp download path. path=" << path.path;
             }
-        } catch (...) {
-            LOG(WARNING) << "std exception when remove dpp download path. path=" << path.path;
         }
-    }
 
 #define BUILD_DYNAMIC_TASK_THREAD_POOL(name, min_threads, max_threads, queue_size, pool) \
     do {                                                                                 \
@@ -144,71 +148,72 @@ void AgentServer::Impl::init_or_die() {
         CHECK(st.ok()) << st;                                                            \
     } while (false)
 
-    // The ideal queue size of threadpool should be larger than the maximum number of tablet of a partition.
-    // But it seems that there's no limit for the number of tablets of a partition.
-    // Since a large queue size brings a little overhead, a big one is chosen here.
-    BUILD_DYNAMIC_TASK_THREAD_POOL("publish_version", config::transaction_publish_version_worker_count,
-                                   config::transaction_publish_version_worker_count,
-                                   DEFAULT_DYNAMIC_THREAD_POOL_QUEUE_SIZE, _thread_pool_publish_version);
+        // The ideal queue size of threadpool should be larger than the maximum number of tablet of a partition.
+        // But it seems that there's no limit for the number of tablets of a partition.
+        // Since a large queue size brings a little overhead, a big one is chosen here.
+        BUILD_DYNAMIC_TASK_THREAD_POOL("publish_version", config::transaction_publish_version_worker_count,
+                                       config::transaction_publish_version_worker_count,
+                                       DEFAULT_DYNAMIC_THREAD_POOL_QUEUE_SIZE, _thread_pool_publish_version);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("drop", config::drop_tablet_worker_count, config::drop_tablet_worker_count,
-                                   std::numeric_limits<int>::max(), _thread_pool_drop);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("drop", config::drop_tablet_worker_count, config::drop_tablet_worker_count,
+                                       std::numeric_limits<int>::max(), _thread_pool_drop);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("create_tablet", config::create_tablet_worker_count,
-                                   config::create_tablet_worker_count, std::numeric_limits<int>::max(),
-                                   _thread_pool_create_tablet);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("create_tablet", config::create_tablet_worker_count,
+                                       config::create_tablet_worker_count, std::numeric_limits<int>::max(),
+                                       _thread_pool_create_tablet);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("alter_tablet", config::alter_tablet_worker_count, config::alter_tablet_worker_count,
-                                   std::numeric_limits<int>::max(), _thread_pool_alter_tablet);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("alter_tablet", config::alter_tablet_worker_count,
+                                       config::alter_tablet_worker_count, std::numeric_limits<int>::max(),
+                                       _thread_pool_alter_tablet);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("clear_transaction", config::clear_transaction_task_worker_count,
-                                   config::clear_transaction_task_worker_count, std::numeric_limits<int>::max(),
-                                   _thread_pool_clear_transaction);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("clear_transaction", config::clear_transaction_task_worker_count,
+                                       config::clear_transaction_task_worker_count, std::numeric_limits<int>::max(),
+                                       _thread_pool_clear_transaction);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("storage_medium_migrate", config::storage_medium_migrate_count,
-                                   config::storage_medium_migrate_count, std::numeric_limits<int>::max(),
-                                   _thread_pool_storage_medium_migrate);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("storage_medium_migrate", config::storage_medium_migrate_count,
+                                       config::storage_medium_migrate_count, std::numeric_limits<int>::max(),
+                                       _thread_pool_storage_medium_migrate);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("check_consistency", config::check_consistency_worker_count,
-                                   config::check_consistency_worker_count, std::numeric_limits<int>::max(),
-                                   _thread_pool_check_consistency);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("check_consistency", config::check_consistency_worker_count,
+                                       config::check_consistency_worker_count, std::numeric_limits<int>::max(),
+                                       _thread_pool_check_consistency);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("upload", config::upload_worker_count, config::upload_worker_count,
-                                   std::numeric_limits<int>::max(), _thread_pool_upload);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("upload", config::upload_worker_count, config::upload_worker_count,
+                                       std::numeric_limits<int>::max(), _thread_pool_upload);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("download", config::download_worker_count, config::download_worker_count,
-                                   std::numeric_limits<int>::max(), _thread_pool_download);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("download", config::download_worker_count, config::download_worker_count,
+                                       std::numeric_limits<int>::max(), _thread_pool_download);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("make_snapshot", config::make_snapshot_worker_count,
-                                   config::make_snapshot_worker_count, std::numeric_limits<int>::max(),
-                                   _thread_pool_make_snapshot);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("make_snapshot", config::make_snapshot_worker_count,
+                                       config::make_snapshot_worker_count, std::numeric_limits<int>::max(),
+                                       _thread_pool_make_snapshot);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("release_snapshot", config::release_snapshot_worker_count,
-                                   config::release_snapshot_worker_count, std::numeric_limits<int>::max(),
-                                   _thread_pool_release_snapshot);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("release_snapshot", config::release_snapshot_worker_count,
+                                       config::release_snapshot_worker_count, std::numeric_limits<int>::max(),
+                                       _thread_pool_release_snapshot);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("move_dir", 1, 1, std::numeric_limits<int>::max(), _thread_pool_move_dir);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("move_dir", 1, 1, std::numeric_limits<int>::max(), _thread_pool_move_dir);
 
-    BUILD_DYNAMIC_TASK_THREAD_POOL("update_tablet_meta_info", 1, 1, std::numeric_limits<int>::max(),
-                                   _thread_pool_update_tablet_meta_info);
+        BUILD_DYNAMIC_TASK_THREAD_POOL("update_tablet_meta_info", 1, 1, std::numeric_limits<int>::max(),
+                                       _thread_pool_update_tablet_meta_info);
 
 #ifndef BE_TEST
-    // Currently FE can have at most num_of_storage_path * schedule_slot_num_per_path(default 2) clone tasks
-    // scheduled simultaneously, but previously we have only 3 clone worker threads by default,
-    // so this is to keep the dop of clone task handling in sync with FE.
-    //
-    // TODO(shangyiming): using dynamic thread pool to handle task directly instead of using TaskThreadPool
-    // Currently, the task submission and processing logic is deeply coupled with TaskThreadPool, change that will
-    // need to modify many interfaces. So for now we still use TaskThreadPool to submit clone tasks, but with
-    // only a single worker thread, then we use dynamic thread pool to handle the task concurrently in clone task
-    // callback, so that we can match the dop of FE clone task scheduling.
-    BUILD_DYNAMIC_TASK_THREAD_POOL("clone", MIN_CLONE_TASK_THREADS_IN_POOL,
-                                   _exec_env->store_paths().size() * config::parallel_clone_task_per_path,
-                                   DEFAULT_DYNAMIC_THREAD_POOL_QUEUE_SIZE, _thread_pool_clone);
+        // Currently FE can have at most num_of_storage_path * schedule_slot_num_per_path(default 2) clone tasks
+        // scheduled simultaneously, but previously we have only 3 clone worker threads by default,
+        // so this is to keep the dop of clone task handling in sync with FE.
+        //
+        // TODO(shangyiming): using dynamic thread pool to handle task directly instead of using TaskThreadPool
+        // Currently, the task submission and processing logic is deeply coupled with TaskThreadPool, change that will
+        // need to modify many interfaces. So for now we still use TaskThreadPool to submit clone tasks, but with
+        // only a single worker thread, then we use dynamic thread pool to handle the task concurrently in clone task
+        // callback, so that we can match the dop of FE clone task scheduling.
+        BUILD_DYNAMIC_TASK_THREAD_POOL("clone", MIN_CLONE_TASK_THREADS_IN_POOL,
+                                       _exec_env->store_paths().size() * config::parallel_clone_task_per_path,
+                                       DEFAULT_DYNAMIC_THREAD_POOL_QUEUE_SIZE, _thread_pool_clone);
 #endif
 
-    // It is the same code to create workers of each type, so we use a macro
-    // to make code to be more readable.
+        // It is the same code to create workers of each type, so we use a macro
+        // to make code to be more readable.
 #ifndef BE_TEST
 #define CREATE_AND_START_POOL(pool_name, CLASS_NAME, worker_num) \
     pool_name.reset(new CLASS_NAME(_exec_env, worker_num));      \
@@ -217,50 +222,53 @@ void AgentServer::Impl::init_or_die() {
 #define CREATE_AND_START_POOL(pool_name, CLASS_NAME, worker_num)
 #endif // BE_TEST
 
-    CREATE_AND_START_POOL(_publish_version_workers, PublishVersionTaskWorkerPool, base::NumCPUs())
-    // Both PUSH and REALTIME_PUSH type use _push_workers
-    CREATE_AND_START_POOL(_push_workers, PushTaskWorkerPool,
-                          config::push_worker_count_high_priority + config::push_worker_count_normal_priority)
-    CREATE_AND_START_POOL(_delete_workers, DeleteTaskWorkerPool,
-                          config::delete_worker_count_normal_priority + config::delete_worker_count_high_priority)
-    CREATE_AND_START_POOL(_report_task_workers, ReportTaskWorkerPool, REPORT_TASK_WORKER_COUNT)
-    CREATE_AND_START_POOL(_report_disk_state_workers, ReportDiskStateTaskWorkerPool, REPORT_DISK_STATE_WORKER_COUNT)
-    CREATE_AND_START_POOL(_report_tablet_workers, ReportOlapTableTaskWorkerPool, REPORT_OLAP_TABLE_WORKER_COUNT)
-    CREATE_AND_START_POOL(_report_workgroup_workers, ReportWorkgroupTaskWorkerPool, REPORT_WORKGROUP_WORKER_COUNT)
+        CREATE_AND_START_POOL(_publish_version_workers, PublishVersionTaskWorkerPool, base::NumCPUs())
+        // Both PUSH and REALTIME_PUSH type use _push_workers
+        CREATE_AND_START_POOL(_push_workers, PushTaskWorkerPool,
+                              config::push_worker_count_high_priority + config::push_worker_count_normal_priority)
+        CREATE_AND_START_POOL(_delete_workers, DeleteTaskWorkerPool,
+                              config::delete_worker_count_normal_priority + config::delete_worker_count_high_priority)
+        CREATE_AND_START_POOL(_report_task_workers, ReportTaskWorkerPool, REPORT_TASK_WORKER_COUNT)
+        CREATE_AND_START_POOL(_report_disk_state_workers, ReportDiskStateTaskWorkerPool, REPORT_DISK_STATE_WORKER_COUNT)
+        CREATE_AND_START_POOL(_report_tablet_workers, ReportOlapTableTaskWorkerPool, REPORT_OLAP_TABLE_WORKER_COUNT)
+        CREATE_AND_START_POOL(_report_workgroup_workers, ReportWorkgroupTaskWorkerPool, REPORT_WORKGROUP_WORKER_COUNT)
+    }
     CREATE_AND_START_POOL(_report_resource_usage_workers, ReportResourceUsageTaskWorkerPool,
                           REPORT_RESOURCE_USAGE_WORKER_COUNT)
 #undef CREATE_AND_START_POOL
 }
 
 void AgentServer::Impl::stop() {
-    _thread_pool_publish_version->shutdown();
-    _thread_pool_drop->shutdown();
-    _thread_pool_create_tablet->shutdown();
-    _thread_pool_alter_tablet->shutdown();
-    _thread_pool_clear_transaction->shutdown();
-    _thread_pool_storage_medium_migrate->shutdown();
-    _thread_pool_check_consistency->shutdown();
-    _thread_pool_upload->shutdown();
-    _thread_pool_download->shutdown();
-    _thread_pool_make_snapshot->shutdown();
-    _thread_pool_release_snapshot->shutdown();
-    _thread_pool_move_dir->shutdown();
-    _thread_pool_update_tablet_meta_info->shutdown();
+    if (!_is_compute_node) {
+        _thread_pool_publish_version->shutdown();
+        _thread_pool_drop->shutdown();
+        _thread_pool_create_tablet->shutdown();
+        _thread_pool_alter_tablet->shutdown();
+        _thread_pool_clear_transaction->shutdown();
+        _thread_pool_storage_medium_migrate->shutdown();
+        _thread_pool_check_consistency->shutdown();
+        _thread_pool_upload->shutdown();
+        _thread_pool_download->shutdown();
+        _thread_pool_make_snapshot->shutdown();
+        _thread_pool_release_snapshot->shutdown();
+        _thread_pool_move_dir->shutdown();
+        _thread_pool_update_tablet_meta_info->shutdown();
 
 #ifndef BE_TEST
-    _thread_pool_clone->shutdown();
+        _thread_pool_clone->shutdown();
 #define STOP_POOL(type, pool_name) pool_name->stop();
 #else
 #define STOP_POOL(type, pool_name)
 #endif // BE_TEST
-    STOP_POOL(PUBLISH_VERSION, _publish_version_workers);
-    // Both PUSH and REALTIME_PUSH type use _push_workers
-    STOP_POOL(PUSH, _push_workers);
-    STOP_POOL(DELETE, _delete_workers);
-    STOP_POOL(REPORT_TASK, _report_task_workers);
-    STOP_POOL(REPORT_DISK_STATE, _report_disk_state_workers);
-    STOP_POOL(REPORT_OLAP_TABLE, _report_tablet_workers);
-    STOP_POOL(REPORT_WORKGROUP, _report_workgroup_workers);
+        STOP_POOL(PUBLISH_VERSION, _publish_version_workers);
+        // Both PUSH and REALTIME_PUSH type use _push_workers
+        STOP_POOL(PUSH, _push_workers);
+        STOP_POOL(DELETE, _delete_workers);
+        STOP_POOL(REPORT_TASK, _report_task_workers);
+        STOP_POOL(REPORT_DISK_STATE, _report_disk_state_workers);
+        STOP_POOL(REPORT_OLAP_TABLE, _report_tablet_workers);
+        STOP_POOL(REPORT_WORKGROUP, _report_workgroup_workers);
+    }
     STOP_POOL(REPORT_WORKGROUP, _report_resource_usage_workers);
 #undef STOP_POOL
 }
@@ -547,7 +555,8 @@ ThreadPool* AgentServer::Impl::get_thread_pool(int type) const {
     return nullptr;
 }
 
-AgentServer::AgentServer(ExecEnv* exec_env) : _impl(std::make_unique<AgentServer::Impl>(exec_env)) {}
+AgentServer::AgentServer(ExecEnv* exec_env, bool is_compute_node)
+        : _impl(std::make_unique<AgentServer::Impl>(exec_env, is_compute_node)) {}
 
 AgentServer::~AgentServer() = default;
